// src/DanskeSpil/KlubLotto/Scripts/Types/sitecore.ts
export type SitecoreFile = {
  alt: string;
  border: string;
  class: string;
  mediaExists: boolean;
  mediaId: string;
  src: string;
  title: string;
  width: number;
  height: number;
}

export type SitecoreGenericLink = {
  anchor: string;
  class: string;
  query: string;
  style: string;
  target: string;
  targetId: string;
  text: string;
  title: string;
  type: number;
  url: string;
}

export type SitecoreImage = {
  alt: string;
  border: string;
  class: string;
  mediaExists: boolean;
  mediaId: string;
  src: string;
  title: string;
  width: number;
  height: number;
}